python_face_service/requirements.txt
Flask==3.0.0
face-recognition==1.3.0
Pillow==10.1.0
numpy==1.24.3
dlib==19.24.2
cmake==3.27.7